Refactoring framework
