+++ modules/elo/elo.py
@@ -1,9 +1,9 @@
 from prettytable import PrettyTable
 from trueskill import Rating
-from app.modules.elo.player import Player
+from modules.elo.player import Player
 import sqlite3
 
-conn = sqlite3.connect('app/modules/elo/db/trueskill.db')
+conn = sqlite3.connect('modules/elo/db/trueskill.db')
 
 def ranks(smash):
     if not is_valid_game(smash):
