+++ modules/poll/poll.py
@@ -2,7 +2,7 @@ from app.utils import helper
 from app.receiver.receiver import Receiver
 from app.receiver import receiver
 from app.mac import mac
-from app.modules.poll.voter import Voter
+from modules.poll.voter import Voter
 
 class WAPoll(Receiver):
     def __init__(self, instance, conversation, creator, title, identifier="#"):
