+++ modules/trip/db/schema/schema.sql
